backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
sqlalchemy==2.0.36
alembic==1.14.0
pyodbc==5.2.0
aioodbc==0.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
apscheduler==3.10.4
python-dateutil==2.9.0.post0
httpx==0.28.1
openpyxl==3.1.5
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
anyio==4.7.0
